backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
firebase-admin==5.2.0
reportlab==4.2.2
werkzeug==2.0.3